Move `component/hooks` into `hooks/`
