+++ x-pack/plugins/security_solution/public/management/components/artifact_card_grid/components/grid_header.tsx
@@ -11,7 +11,7 @@ import { CommonProps, EuiText, EuiButtonIcon } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
 import styled from 'styled-components';
 import { CardCompressedHeaderLayout, CardSectionPanel } from '../../artifact_entry_card';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 
 const GridHeaderContainer = styled(CardSectionPanel)`
   padding-top: 0;
