+++ x-pack/plugins/security_solution/public/management/components/artifact_entry_card/components/card_header.tsx
@@ -9,7 +9,7 @@ import React, { memo } from 'react';
 import { CommonProps, EuiFlexGroup, EuiFlexItem, EuiTitle } from '@elastic/eui';
 import styled from 'styled-components';
 import { DateFieldValue } from './date_field_value';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 import { CardActionsFlexItem, CardActionsFlexItemProps } from './card_actions_flex_item';
 
 export interface CardHeaderProps
