+++ x-pack/plugins/security_solution/public/management/components/artifact_entry_card/components/card_sub_header.tsx
@@ -9,7 +9,7 @@ import React, { memo } from 'react';
 import { CommonProps, EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import { TouchedByUsers, TouchedByUsersProps } from './touched_by_users';
 import { EffectScope, EffectScopeProps } from './effect_scope';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 
 export type SubHeaderProps = TouchedByUsersProps &
   EffectScopeProps &
