+++ x-pack/plugins/security_solution/public/management/components/artifact_entry_card/components/date_field_value.tsx
@@ -14,7 +14,7 @@ import {
   FormattedRelativePreferenceDateProps,
 } from '../../../../common/components/formatted_date';
 import { TextValueDisplay } from './text_value_display';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 
 const StyledEuiFlexItem = styled(EuiFlexItem)`
   padding-top: 2px;
