+++ x-pack/plugins/security_solution/public/management/components/artifact_entry_card/components/effect_scope.tsx
@@ -17,7 +17,7 @@ import {
 import { TextValueDisplay } from './text_value_display';
 import { ContextMenuWithRouterSupport } from '../../context_menu_with_router_support';
 import { ContextMenuItemNavByRouterProps } from '../../context_menu_with_router_support/context_menu_item_nav_by_router';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 
 // FIXME:PT support being able to show per policy label for Artifacst that have >0 policies, but no menu
 //          the intent in this component was to also support to be able to display only text for artifacts
