+++ x-pack/plugins/security_solution/public/management/components/artifact_entry_card/components/touched_by_users.tsx
@@ -10,7 +10,7 @@ import { CommonProps, EuiAvatar, EuiBadge, EuiFlexGroup, EuiFlexItem } from '@el
 import styled from 'styled-components';
 import { CREATED_BY, LAST_UPDATED_BY } from './translations';
 import { TextValueDisplay } from './text_value_display';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 
 const StyledEuiFlexItem = styled(EuiFlexItem)`
   margin: 6px;
