+++ x-pack/plugins/security_solution/public/management/components/artifact_list_page/components/artifact_delete_modal.tsx
@@ -20,7 +20,7 @@ import {
   EuiText,
 } from '@elastic/eui';
 import { AutoFocusButton } from '../../../../common/components/autofocus_button/autofocus_button';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 import {
   getPolicyIdsFromArtifact,
   isArtifactGlobal,
