+++ x-pack/plugins/security_solution/public/management/components/artifact_list_page/components/artifact_flyout.tsx
@@ -24,9 +24,9 @@ import {
 
 import { EuiFlyoutSize } from '@elastic/eui/src/components/flyout/flyout';
 import { HttpFetchError } from '@kbn/core/public';
-import { useUrlParams } from '../../hooks/use_url_params';
+import { useUrlParams } from '../../../hooks/use_url_params';
 import { useIsFlyoutOpened } from '../hooks/use_is_flyout_opened';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 import { useSetUrlParams } from '../hooks/use_set_url_params';
 import {
   ArtifactFormComponentOnChangeCallbackProps,
@@ -39,7 +39,7 @@ import { useKibana, useToasts } from '../../../../common/lib/kibana';
 import { createExceptionListItemForCreate } from '../../../../../common/endpoint/service/artifacts/utils';
 import { useWithArtifactSubmitData } from '../hooks/use_with_artifact_submit_data';
 import { useIsArtifactAllowedPerPolicyUsage } from '../hooks/use_is_artifact_allowed_per_policy_usage';
-import { useIsMounted } from '../../hooks/use_is_mounted';
+import { useIsMounted } from '../../../hooks/use_is_mounted';
 import { useGetArtifact } from '../../../hooks/artifacts';
 import type { PolicyData } from '../../../../../common/endpoint/types';
