+++ x-pack/plugins/security_solution/public/management/components/artifact_list_page/hooks/use_is_flyout_opened.ts
@@ -6,7 +6,7 @@
  */
 
 import { useMemo } from 'react';
-import { useUrlParams } from '../../hooks/use_url_params';
+import { useUrlParams } from '../../../hooks/use_url_params';
 import { ArtifactListPageUrlParams } from '../types';
 
 const SHOW_VALUES: readonly string[] = ['create', 'edit'];
