+++ x-pack/plugins/security_solution/public/management/components/artifact_list_page/hooks/use_with_artifact_list_data.ts
@@ -9,15 +9,15 @@ import { useEffect, useMemo, useState } from 'react';
 import { Pagination } from '@elastic/eui';
 import { useQuery } from 'react-query';
 import type { ServerApiError } from '../../../../common/types';
-import { useIsMounted } from '../../hooks/use_is_mounted';
+import { useIsMounted } from '../../../hooks/use_is_mounted';
 import { MANAGEMENT_PAGE_SIZE_OPTIONS } from '../../../common/constants';
-import { useUrlParams } from '../../hooks/use_url_params';
+import { useUrlParams } from '../../../hooks/use_url_params';
 import { ExceptionsListApiClient } from '../../../services/exceptions_list/exceptions_list_api_client';
 import { ArtifactListPageUrlParams } from '../types';
 import { MaybeImmutable } from '../../../../../common/endpoint/types';
 import { useKueryFromExceptionsSearchFilter } from './use_kuery_from_exceptions_search_filter';
 import { useListArtifact } from '../../../hooks/artifacts';
-import { useUrlPagination } from '../../hooks/use_url_pagination';
+import { useUrlPagination } from '../../../hooks/use_url_pagination';
 
 type WithArtifactListDataInterface = ReturnType<typeof useListArtifact> & {
   /**
