+++ x-pack/plugins/security_solution/public/management/components/console/components/command_input/command_input.tsx
@@ -11,7 +11,7 @@ import styled from 'styled-components';
 import classNames from 'classnames';
 import { KeyCapture, KeyCaptureProps } from './key_capture';
 import { useConsoleStateDispatch } from '../../hooks/state_selectors/use_console_state_dispatch';
-import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../../hooks/use_test_id_generator';
 import { useDataTestSubj } from '../../hooks/state_selectors/use_data_test_subj';
 
 const CommandInputContainer = styled.div`
