+++ x-pack/plugins/security_solution/public/management/components/console/components/command_input/key_capture.tsx
@@ -16,7 +16,7 @@ import React, {
 } from 'react';
 import { pick } from 'lodash';
 import styled from 'styled-components';
-import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../../hooks/use_test_id_generator';
 import { useDataTestSubj } from '../../hooks/state_selectors/use_data_test_subj';
 
 const NOOP = () => undefined;
