+++ x-pack/plugins/security_solution/public/management/components/console/components/command_list.tsx
@@ -9,7 +9,7 @@ import React, { memo, useMemo } from 'react';
 import { EuiDescriptionList, EuiFlexGroup, EuiFlexItem, EuiText } from '@elastic/eui';
 import { FormattedMessage } from '@kbn/i18n-react';
 import { CommandDefinition } from '../types';
-import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../hooks/use_test_id_generator';
 import { useDataTestSubj } from '../hooks/state_selectors/use_data_test_subj';
 
 export interface CommandListProps {
