+++ x-pack/plugins/security_solution/public/management/components/context_menu_with_router_support/context_menu_item_nav_by_router.tsx
@@ -15,7 +15,7 @@ import {
 import styled from 'styled-components';
 import { NavigateToAppOptions } from '@kbn/core/public';
 import { useNavigateToAppEventHandler } from '../../../common/hooks/endpoint/use_navigate_to_app_event_handler';
-import { useTestIdGenerator } from '../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
 
 export interface ContextMenuItemNavByRouterProps extends EuiContextMenuItemProps {
   /** The Kibana (plugin) app id */
