+++ x-pack/plugins/security_solution/public/management/components/context_menu_with_router_support/context_menu_with_router_support.tsx
@@ -20,7 +20,7 @@ import {
   ContextMenuItemNavByRouter,
   ContextMenuItemNavByRouterProps,
 } from './context_menu_item_nav_by_router';
-import { useTestIdGenerator } from '../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
 
 export interface ContextMenuWithRouterSupportProps
   extends CommonProps,
