+++ x-pack/plugins/security_solution/public/management/components/effected_policy_select/effected_policy_select.tsx
@@ -26,7 +26,7 @@ import styled from 'styled-components';
 import { PolicyData } from '../../../../common/endpoint/types';
 import { LinkToApp } from '../../../common/components/endpoint/link_to_app';
 import { getPolicyDetailPath } from '../../common/routing';
-import { useTestIdGenerator } from '../hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../hooks/use_test_id_generator';
 import { useAppUrl } from '../../../common/lib/kibana/hooks';
 import { Loader } from '../../../common/components/loader';
