+++ x-pack/plugins/security_solution/public/management/hooks/use_is_mounted.ts
