+++ x-pack/plugins/security_solution/public/management/hooks/use_url_pagination.ts
@@ -7,7 +7,7 @@
 
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
-import { MANAGEMENT_DEFAULT_PAGE_SIZE, MANAGEMENT_PAGE_SIZE_OPTIONS } from '../../common/constants';
+import { MANAGEMENT_DEFAULT_PAGE_SIZE, MANAGEMENT_PAGE_SIZE_OPTIONS } from '../common/constants';
 import { useUrlParams } from './use_url_params';
 
 // Page is not index-based, it is 1-based
