+++ x-pack/plugins/security_solution/public/management/hooks/use_url_params.ts
