+++ x-pack/plugins/security_solution/public/management/pages/policy/view/artifacts/layout/policy_artifacts_layout.tsx
@@ -21,7 +21,7 @@ import { useAppUrl } from '../../../../../../common/lib/kibana';
 import { APP_UI_ID } from '../../../../../../../common/constants';
 import { ImmutableObject, PolicyData } from '../../../../../../../common/endpoint/types';
 import { ManagementPageLoader } from '../../../../../components/management_page_loader';
-import { useUrlParams } from '../../../../../components/hooks/use_url_params';
+import { useUrlParams } from '../../../../../hooks/use_url_params';
 import { useUserPrivileges } from '../../../../../../common/components/user_privileges';
 import { usePolicyDetailsArtifactsNavigateCallback } from '../../policy_hooks';
 import { ExceptionsListApiClient } from '../../../../../services/exceptions_list/exceptions_list_api_client';
