+++ x-pack/plugins/security_solution/public/management/pages/policy/view/ingest_manager_integration/endpoint_package_custom_extension/components/fleet_artifacts_card.tsx
@@ -19,7 +19,7 @@ import { ExceptionItemsSummary } from './exception_items_summary';
 import { StyledEuiFlexGridGroup, StyledEuiFlexGridItem } from './styled_components';
 import { useSummaryArtifact } from '../../../../../../hooks/artifacts';
 import { ExceptionsListApiClient } from '../../../../../../services/exceptions_list/exceptions_list_api_client';
-import { useTestIdGenerator } from '../../../../../../components/hooks/use_test_id_generator';
+import { useTestIdGenerator } from '../../../../../../hooks/use_test_id_generator';
 
 const ARTIFACTS_LABELS = {
   artifactsSummaryApiError: (error: string) =>
